[refactory] api.js 호출 위치 변경
