+++ src/main.js
@@ -2,7 +2,7 @@
 import './css/style.css'
 
 // JS 호출
-import '../api/api.js'
+import './api/api.js'
 import './js/components/_popup.js'
 import './js/components/_share.js'
 import './js/components/_phraseLoader.js'
