+++ src/pages/result/result.js
@@ -17,7 +17,7 @@ import {
   handleShowResult,
 } from '../../js/components/_imageLoading.js'
 import { shareResult } from '../../js/components/_share.js'
-import { getData, getUser } from '../../../api/api.js'
+import { getData, getUser } from '../../api/api.js'
 import { updateUserDiSplay } from '../../js/components/_popup.js'
 
 /**
